Add Todos for using jenkins AMI builded with packer
The current implementation of the jenkins master VM creation use always
the ubuntu base image and all the installation steps for jenkins handled
with cloud init.

This setup works but its takes couple of minutes and can result in an
non working jenkins m,aster because of cloud init failures like
temporarly apt repository outages for example.

The next step is to check if there is an jenkins master AMI available
and use this one without cloud-init.